# docs/swagger/docs.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "email": "hi@fanzru.dev",
      "name": "Social Media Service Team"
    },
    "description": "API for social media service",
    "title": "Social Media Service API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/comments/user/{userId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "User ID",
            "format": "int64",
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Cursor for pagination",
            "in": "query",
            "name": "cursor",
            "required": false,
            "type": "string"
          },
          {
            "default": 20,
            "description": "Number of comments to return (max 100)",
            "in": "query",
            "maximum": 100,
            "minimum": 1,
            "name": "limit",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "User comments retrieved successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "tags": [
          "Comments"
        ],
        "description": "Get all comments by a specific user",
        "summary": "Get user comments"
      }
    },
    "/api/comments/{id}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Comment ID",
            "format": "int64",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "403": {
            "description": "Forbidden - not the comment creator",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "404": {
            "description": "Comment not found",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Comments"
        ],
        "description": "Delete a comment (only the creator can delete)",
        "summary": "Delete comment"
      },
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Comment ID",
            "format": "int64",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment retrieved successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "404": {
            "description": "Comment not found",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "tags": [
          "Comments"
        ],
        "description": "Get a specific comment by its ID",
        "summary": "Get comment by ID"
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Comment ID",
            "format": "int64",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCommentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment updated successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "400": {
            "description": "Bad request - validation errors",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "403": {
            "description": "Forbidden - not the comment creator",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "404": {
            "description": "Comment not found",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Comments"
        ],
        "description": "Update a comment (only the creator can update)",
        "summary": "Update comment"
      }
    },
    "/api/posts/{postId}/comments": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Post ID",
            "format": "int64",
            "in": "path",
            "name": "postId",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Cursor for pagination",
            "in": "query",
            "name": "cursor",
            "required": false,
            "type": "string"
          },
          {
            "default": 20,
            "description": "Number of comments to return (max 100)",
            "in": "query",
            "maximum": 100,
            "minimum": 1,
            "name": "limit",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "404": {
            "description": "Post not found",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "tags": [
          "Comments"
        ],
        "description": "Get all comments for a specific post with pagination",
        "summary": "Get post comments"
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Post ID",
            "format": "int64",
            "in": "path",
            "name": "postId",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCommentRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "400": {
            "description": "Bad request - validation errors",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "404": {
            "description": "Post not found",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Comments"
        ],
        "description": "Create a new comment on a specific post",
        "summary": "Create a new comment"
      }
    },
...
  },
  "definitions": {
...
    "StandardResponse": {
      "properties": {
        "code": {
          "enum": [
            "SUCCESS",
            "FAILED",
            "BAD_REQUEST",
            "UNAUTHORIZED",
            "FORBIDDEN",
            "NOT_FOUND",
            "CONFLICT",
            "INTERNAL_SERVER_ERROR"
          ],
          "example": "SUCCESS",
          "type": "string"
        },
        "data": {
          "description": "Response data (varies by endpoint)",
          "type": "object"
        },
        "errors": {
          "example": [],
          "items": {
            "type": "string"
          },
          "type": "array"
        },
        "message": {
          "example": "Operation completed successfully",
          "type": "string"
        },
        "requestId": {
          "example": "req_123456789",
          "type": "string"
        },
        "serverTime": {
          "example": "2024-01-01T00:00:00Z",
          "format": "date-time",
          "type": "string"
        }
      },
      "type": "object"
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "description": "JWT token obtained from login endpoint",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
...
}